Made requested changes
