+++ client/src/components/overview/ProductInfo.jsx
@@ -3,17 +3,14 @@ import {GrFacebook} from 'react-icons/gr';
 import { AiFillTwitterSquare } from 'react-icons/ai';
 import { FaPinterestSquare } from 'react-icons/fa';
 import {FacebookShareButton, TwitterShareButton, PinterestShareButton} from 'react-share';
+import CrossedOutPrice from '../CrossedOutPrice';
 
 function ProductInfo (props) {
   const renderPrice = () => {
     if (props.salePrice === null) {
       return <p>{props.price}</p>;
     } else {
-      return(
-        <div>
-          <p className="crossed-out">{props.price}</p>
-          <p>{props.salePrice}</p>
-        </div>);
+      return(<CrossedOutPrice price={props.price} salePrice={props.salePrice}/>);
 
     }
   };
@@ -28,22 +25,22 @@ function ProductInfo (props) {
         {/*NEED TO CHANGE URLS TO THE CORRECT ONE*/}
         <span className="share">
           <FacebookShareButton
-            url="http://localhost:3000/"
+            url="#"
             quote="Buy this instantly">
             <GrFacebook/>
           </FacebookShareButton>
         </span>
         <span className="share">
           <TwitterShareButton
-            url="http://localhost:3000/"
+            url="#"
             title="Haworthia">
             <AiFillTwitterSquare/>
           </TwitterShareButton>
         </span>
         <span className="share">
           <PinterestShareButton
-            url="http://localhost:3000/"
-            media="http://github.com"
+            url="#"
+            media="#"
             description='This piece is dope'>
             <FaPinterestSquare/>
           </PinterestShareButton>
